Rename integration test scripts
