+++ tests/smallbutchallenging/generate-reference.py
@@ -1,0 +1,1 @@
+../litmus/generate-reference.py
