--- tests/smallbutchallenging/test-swsc.py
@@ -1,1 +1,0 @@
-../litmus/test-swsc.py
